refactor of entrypoints and files.
