+++ README.md
@@ -6,8 +6,9 @@ The over-powered LED clock for raspberry pis and the TM1637 LED Display.
 
 LED-Kurokku is a versatile system for controlling TM1637 LED displays with multiple widgets, automated brightness control, and weather integration. The system consists of:
 
-1. **led-kurokku**: Core application that controls the LED display
-2. **kurokku-cli**: Command-line tool for managing multiple LED-Kurokku instances as well as launching a **web interface** for visualizing the TM1637 display in a browser.
+1. **led-kurokku**: Core application that controls the LED display or a virtual text display.
+2. **web-kurokku**: Web server for running a kurokku instance as a web page (essentially a web simulator of a TM1637 display).
+3. **kurokku-cli**: Command-line tool for managing multiple LED-Kurokku instances and weather configuration and weather server.
 
 See the [Architecture Document](ARCHITECTURE.md) for a detailed system diagram and explanation of how the components work together.
 
@@ -16,6 +17,8 @@ The clock can be configured to show various widgets at specific intervals.
 Widget types:
 * clock
   * configurable to show either 12/24 hour format.
+    * when in 12 hour format (the colon blinks double while PM and single blink during AM).
+  * 
 * message
   * these can be static messages or dynamically loaded on display from a Redis compatible store. So if you have a separate process updating a temperature reading in redis, you can have a static message that displays that reading. 
 * alert
@@ -47,7 +50,7 @@ If you don't install the optional `rpi` dependencies or they're not installed co
 You can run LED-Kurokku with a web-based virtual display that shows the TM1637 segments in a browser:
 
 ```bash
-kurokku-cli web start
+web-kurokku
 ```
 
 This will start a web server on port 8080 (by default) with these features:
@@ -120,11 +123,47 @@ Or you can choose to use the virtual or real display but also log debug level ou
 
 ### Example Configuration
 
-*TODO:* example configuration
+The following example configures:
+- widgets
+  - alert widget - only displays if there is an alert.
+  - clock - the time (in 12 hour format) - for 5 seconds
+  - message - dynamically fetching a temperature message from the local redis instance and displaying for 5 seconds
+- brightness
+  - high brightness (7) starting at 5:40:48
+  - low brightness (5) starting at 19:47:35 (7:47:35 PM)
+
+Using the cli to run the weather server, it will update the brightness begin and end times based on sunrise and sunset for the default weather location for each configured kurokku instance.
+
+```YAML
+widgets:
+- widget_type: alert
+  enabled: true
+  duration: 0
+  scroll_speed: 0.15
+  repeat: true
+  sleep_before_repeat: 1.0
+- widget_type: clock
+  enabled: true
+  duration: 5
+  use_24_hour_format: false
+- widget_type: message
+  enabled: true
+  duration: 5
+  message: --*F
+  dynamic_source: kurokku:weather:temp:nashville
+  scroll_speed: 0.1
+  repeat: false
+  sleep_before_repeat: 1.0
+brightness:
+  begin: 05:40:48
+  end: '19:47:35'
+  high: 7
+  low: 5
+```
 
 ## CLI Tool
 
-The `kurokku-cli` tool provides a way to manage multiple LED-Kurokku instances. See [README-cli.md](README-cli.md) for detailed usage instructions.
+The `kurokku-cli` tool provides a way to manage multiple LED-Kurokku instances. See [README-cli.md](docs/README-cli.md) for detailed usage instructions.
 
 Key features:
 - Instance management for multiple LED-Kurokku displays
