+++ src/led_kurokku/web/static/README.md
@@ -1,0 +1,1 @@
+static files go here...
